requirements.txt
aiogram==2.25.1
motor==3.3.1
aiohttp==3.8.5
python-dotenv==1.0.0